% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aloha_message.R
\name{aloha_message}
\alias{aloha_message}
\title{Say Aloha}
\usage{
aloha_message(name)
}
\arguments{
\item{name}{(character) A name to say aloha to.}
}
\description{
This function will say aloha to any inputted name.
}
