% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greetings.R
\docType{package}
\name{greetings-package}
\alias{greetings}
\alias{greetings-package}
\title{greetings}
\description{
A package that provides a pleasant way to say hello or goodbye to a friend.
}
\section{greetings functions}{

say_aloha
}

\author{
\strong{Maintainer}: Mitchell Maier \email{mmaier@ucsb.edu}

Authors:
\itemize{
  \item Irene Steves (https://github.com/isteves)
}

}
