% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_aloha.R
\name{say_aloha}
\alias{say_aloha}
\title{Say Aloha}
\usage{
say_aloha(name, print = TRUE)
}
\arguments{
\item{name}{(character) A name to say aloha to.}

\item{print}{(logical) Option to print your message. Defaults to \code{TRUE}}
}
\value{
(character) An aloha message
}
\description{
This function will say aloha to any inputted name.
}
\examples{
# Say hello to a friend
friend <- "Irene"
say_aloha(friend)

}
